naming-update-script/pyproject.toml
[tool.poetry]
name = "naming-update-script"
version = "0.1.0"
description = ""
authors = ["naz"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
python-dotenv = "^0.19.0"
python-logging-loki = "^0.3.1"
ipfshttpclient = {git = "https://github.com/nazariyv/py-ipfs-http-client.git", rev = "9bf6bb2"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
